requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-bdd==7.0.0
pytest-html==4.1.1
allure-pytest==2.13.2
